Concurrent vs Serial requests timings
From our conversation about doing concurrent versus serial
requests, and what the timings look like between spawning a Task
asynchronously to run concurrently for each web request, and doing
a simple map to kick off requests serially.

I have added in calls to `:timer.tc` which will return a tuple
with the first item being the time in *micro* seconds it took for
the operation to complete, and the second is the result (which
is just being thrown away since it is being outputed via IO.puts).

I have included the timings of just kicking off N number of tasks
as well that just return the value `1` for a baseline to see how
long it takes for those to be spawned up and completed.